+++ cloud_export_to_parquet/create_schedule.py
@@ -1,11 +1,7 @@
-"""Module defines run temporal workflow."""
-
 import asyncio
 import traceback
 from datetime import datetime, timedelta
 
-from dataobject import ProtoToParquetWorkflowInput
-from shared import DATA_TRANSFORMATION_TASK_QUEUE_NAME, WORKFLOW_ID_PREFIX
 from temporalio.client import (
     Client,
     Schedule,
@@ -16,12 +12,14 @@ from temporalio.client import (
 )
 from workflows import ProtoToParquet
 
+from cloud_export_to_parquet.workflows import ProtoToParquetWorkflowInput
+
 
 async def main() -> None:
     """Main function to run temporal workflow."""
     # Create client connected to server at the given address
-    client: Client = await Client.connect("localhost:7233", namespace="default")
-    # TODO: update s3_bucket and namespace to the actual name
+    client = await Client.connect("localhost:7233")
+    # TODO: update s3_bucket and namespace to the actual usecase
     wf_input = ProtoToParquetWorkflowInput(
         num_delay_hour=2,
         export_s3_bucket="test-input-bucket",
@@ -29,6 +27,18 @@ async def main() -> None:
         output_s3_bucket="test-output-bucket",
     )
 
+    # Run the workflow
+    # try:
+    #     await client.start_workflow(
+    #         ProtoToParquet.run,
+    #         wf_input,
+    #         id = f"proto-to-parquet-{datetime.now()}",
+    #         task_queue="DATA_TRANSFORMATION_TASK_QUEUE",
+    #     )
+    # except WorkflowFailureError:
+    #     print("Got exception: ", traceback.format_exc())
+
+    # Create the schedule
     try:
         await client.create_schedule(
             "hourly-proto-to-parquet-wf-schedule",
@@ -36,8 +46,8 @@ async def main() -> None:
                 action=ScheduleActionStartWorkflow(
                     ProtoToParquet.run,
                     wf_input,
-                    id=f"{WORKFLOW_ID_PREFIX}-{datetime.now()}",
-                    task_queue=DATA_TRANSFORMATION_TASK_QUEUE_NAME,
+                    id=f"proto-to-parquet-{datetime.now()}",
+                    task_queue="DATA_TRANSFORMATION_TASK_QUEUE",
                 ),
                 spec=ScheduleSpec(
                     intervals=[ScheduleIntervalSpec(every=timedelta(hours=1))]
