--- export_proto_to_parquet/shared.py
@@ -1,2 +1,0 @@
-DATA_TRANSFORMATION_TASK_QUEUE_NAME = "DATA_TRANSFORMATION_TASK_QUEUE"
-WORKFLOW_ID_PREFIX = "proto-to-parquet"
