+++ pyproject.toml
@@ -16,8 +16,11 @@ packages = [
 "Bug Tracker" = "https://github.com/temporalio/samples-python/issues"
 
 [tool.poetry.dependencies]
-python = "^3.8"
+python = "^3.9"
 temporalio = "^1.5.0"
+pandas = "^2.2.2"
+pyarrow = "^15.0.2"
+aioboto3 = "^12.4.0"
 
 [tool.poetry.dev-dependencies]
 black = "^22.3.0"
@@ -51,6 +54,10 @@ openai = "^1.4.0"
 fastapi = "^0.105.0"
 tqdm = "^4.62.0"
 uvicorn = { version = "^0.24.0.post1", extras = ["standard"]}
+
+[tool.poetry.group.cloud_export_to_parquet]
+optional = true
+dependencies = { pyarrow = "^15.0.2", pandas = "^2.2.2", aioboto3 = "^12.4.0"}
 
 [tool.poetry.group.open_telemetry]
 optional = true
